% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricklink-blink_colour_guide.R
\name{blink_colour_guide}
\alias{blink_colour_guide}
\title{Bricklink Colour Guide/Catalogue}
\usage{
blink_colour_guide()
}
\value{
A dataframe with colour information.
}
\description{
Scrape the colour guide from \href{https://www.bricklink.com/catalogColors.asp}{Bricklink} which includes information pertaining to number of parts, number of sets, wanted, quantities for sale, year start and year end.
}
\examples{
\dontrun{
blink_colour_guide()
}
}
