% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickset-bset_import_inventory.R
\name{bset_import_inventory}
\alias{bset_import_inventory}
\title{Brickset Inventory}
\usage{
bset_import_inventory(type = "sets", f_path = "local")
}
\arguments{
\item{type}{A character string specifying the inventory to import. One of \code{"sets"} or \code{"parts"}.}

\item{f_path}{A character string specifying the file path to the inventory csv.}
}
\value{
A dataframe with Lego set information.
}
\description{
Import sets or parts inventory from csv exported from \href{https://brickset.com/export}{brickset}.
}
\examples{
\dontrun{
bset_import_inventory(type = "sets", f_path = "local")
}
}
