% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bset_import_inventory.R
\name{bset_import_sets_inventory}
\alias{bset_import_sets_inventory}
\title{Brickset Sets Inventory}
\usage{
bset_import_sets_inventory(f_path = "local")
}
\arguments{
\item{f_path}{A character string specifying the file path to the inventory csv.}
}
\value{
A dataframe with Lego set information.
}
\description{
Import sets inventory from csv exported from \href{https://brickset.com/export}{brickset}.
}
\examples{
\dontrun{
bset_import_sets_inventory()
}
}
\keyword{internal}
