% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bset_set_themes.R
\name{bset_set_themes}
\alias{bset_set_themes}
\title{Bricket Lego Themes}
\usage{
bset_set_themes()
}
\value{
A dataframe with themes and link to corresponding brickset webpage.
}
\description{
Scrape list of Lego themes from \href{https://brickset.com/browse/sets}{brickset}.
}
\examples{
\dontrun{
bset_set_themes()
}
}
