% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ensureColumn}
\alias{ensureColumn}
\title{Ensure a column exists in the dataset}
\usage{
ensureColumn(dataset, oldName, name)
}
\arguments{
\item{dataset}{the dataset.}

\item{oldName}{the old column name.}

\item{name}{the new column name.}
}
\description{
Ensure a column exists in the dataset
}
