% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink_price_guide.R
\name{extract_for_sale}
\alias{extract_for_sale}
\title{Current Items for Sale}
\usage{
extract_for_sale(...)
}
\arguments{
\item{raw_html}{An list object of class \code{xml_document} imported by \code{rvest::read_html()}.}

\item{type}{A character string specifying the type of listing. Either \code{"new"} or \code{"used"}.}
}
\value{
A data frame with three columns containing price and quantity of used items for sale.
}
\description{
New or used items currently for sale scraped from the bricklink price guide.
}
\keyword{internal}
