% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricklink_price_guide.R
\name{extract_sold}
\alias{extract_sold}
\title{Previously Sold Items}
\usage{
extract_sold(...)
}
\arguments{
\item{sold_raw}{A dataframe with raw text extracted from the \code{xml_document} imported by \code{rvest::read_html()}.}

\item{all_months}{A character string with month labels for regex.}

\item{type}{A character string specifying the type of listing. Either \code{"new"} or \code{"used"}.}
}
\value{
A dataframe with three columns containing price and quantity of new or used items sold in the past six months.
}
\description{
Previously sold new or used items scraped from the bricklink price guide.
}
\keyword{internal}
