% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{filterByYear}
\alias{filterByYear}
\title{Filter dataset only by year}
\usage{
filterByYear(dataset, minYear, maxYear)
}
\arguments{
\item{dataset}{the dataset.}

\item{minYear}{the minimum year.}

\item{maxYear}{the maximum year.}
}
\value{
the filtered dataset with the filters applied.
}
\description{
Filter dataset only by year
}
