% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forex_exchange_rate}
\alias{forex_exchange_rate}
\title{Foreign Exchange Rate Conversion}
\usage{
forex_exchange_rate(
  from = "CAD",
  amount = 1,
  to = c("USD", "EUR", "GBP"),
  as_df = TRUE
)
}
\arguments{
\item{from}{A character string. Three-letter currency abbreviation.}

\item{amount}{A numerical scalae. The amount to be converted.}

\item{to}{A character string or vector. Three-letter currency abbreviation(s).}

\item{as_df}{A logical scalar. Return as dataframe or vector (if \code{FALSE}).}
}
\value{
A dataframe or vector of currency equivalents.
}
\description{
See \href{https://exchangerate.host/#/#docs}{here} for documentation.
}
\keyword{internal}
