% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ggplotCountByYear}
\alias{ggplotCountByYear}
\title{Ggplot count by year}
\usage{
ggplotCountByYear(dataset, title = "Number per Year",
  xAxisLabel = "Year", yAxisLabel = "Number",
  caption = "Source: Darren Redmond's owned dataset")
}
\arguments{
\item{dataset}{the dataset.}

\item{title}{the title.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - the count.}

\item{caption}{the caption.}
}
\value{
the count by year plot.
}
\description{
Ggplot count by year
}
