% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ggplotLego}
\alias{ggplotLego}
\title{ggplot - wrapper for all of the ggplot functionality.}
\usage{
ggplotLego(dataset, type)
}
\arguments{
\item{dataset}{the dataset.}

\item{type}{the type of ggplot to create - can be one of 'year', 'piece', 'price', 'theme', 'piece_theme_avg'.}
}
\value{
the plot.
}
\description{
ggplot - wrapper for all of the ggplot functionality.
}
\examples{
dataset <- read.lego(system.file('extdata', 'brickset-mysets-owned.csv', package ='lego'))
ggplotLego(dataset, 'year')
ggplotLego(dataset, 'piece')
ggplotLego(dataset, 'price')
ggplotLego(dataset, 'theme')
ggplotLego(dataset, 'piece_theme_avg')
}
\keyword{lego}
