% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ggplotPiecesCountByYear}
\alias{ggplotPiecesCountByYear}
\title{Ggplot number of pieces by year}
\usage{
ggplotPiecesCountByYear(dataset, title = "Number of Pieces per Year",
  xAxisLabel = "Year", yAxisLabel = "Number of Pieces",
  caption = "Source: Darren Redmond's owned dataset")
}
\arguments{
\item{dataset}{the dataset.}

\item{title}{the title.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - the number of pieces.}

\item{caption}{the caption.}
}
\value{
the setsByPiecesCount plot.
}
\description{
Ggplot number of pieces by year
}
\examples{
dataset <- read.lego(system.file('extdata', 'brickset-mysets-owned.csv', package ='lego'))
ggplotPiecesCountByYear(groupByPiece(dataset))
ggplotPiecesCountByYear(groupByPiece(dataset, themes=c('Star Wars')))
ggplotPiecesCountByYear(groupByPiece(dataset, 2010, 2016, 5, 10, 100, 1000,
    c('Star Wars'), c('Episode I')))
ggplotPiecesCountByYear(groupByPiece(dataset, 2014, 2015,
    themes=c('Star Wars'), subthemes=c('Episode I')))
}
