% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ggplotPriceByYear}
\alias{ggplotPriceByYear}
\title{Ggplot price by year}
\usage{
ggplotPriceByYear(dataset, title = "Price per Year",
  xAxisLabel = "Year", yAxisLabel = "Price",
  caption = "Source: Darren Redmond's owned dataset")
}
\arguments{
\item{dataset}{the dataset.}

\item{title}{the title.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - the price.}

\item{caption}{the caption.}
}
\value{
the setsByPrice plot.
}
\description{
Ggplot price by year
}
