% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ggplotSetsCountByYear}
\alias{ggplotSetsCountByYear}
\title{Ggplot number of sets by year}
\usage{
ggplotSetsCountByYear(dataset, title = "Number of Sets per Year",
  xAxisLabel = "Year", yAxisLabel = "Number of Sets",
  caption = "Source: Darren Redmond's owned dataset")
}
\arguments{
\item{dataset}{the dataset.}

\item{title}{the title.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - the number of sets.}

\item{caption}{the caption.}
}
\value{
the setsByYear plot.
}
\description{
Ggplot number of sets by year
}
\examples{
dataset <- read.lego(system.file('extdata', 'brickset-mysets-owned.csv', package ='lego'))
ggplotSetsCountByYear(groupByYear(dataset))
ggplotSetsCountByYear(groupByYear(dataset, themes=c('Star Wars')))
ggplotSetsCountByYear(groupByYear(dataset, 2010, 2016, 5, 10, 100, 1000,
    c('Star Wars'), c('Episode I')))
ggplotSetsCountByYear(groupByYear(dataset, 2014, 2015,
    themes=c('Star Wars'), subthemes=c('Episode I')))
}
