% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ggplotThemesCountByYear}
\alias{ggplotThemesCountByYear}
\title{Ggplot number of themes by year}
\usage{
ggplotThemesCountByYear(dataset, title = "Number of Themes per Year",
  xAxisLabel = "Year", yAxisLabel = "Number of Themes",
  caption = "Source: Darren Redmond's owned dataset")
}
\arguments{
\item{dataset}{the dataset.}

\item{title}{the title.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - the price.}

\item{caption}{the caption.}
}
\value{
the themesByYear plot.
}
\description{
Ggplot number of themes by year
}
