% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{groupByPiece}
\alias{groupByPiece}
\title{Group by year to get total count of pieces.}
\usage{
groupByPiece(dataset, minYear = min(dataset$year),
  maxYear = max(dataset$year), minPrice = min(dataset$price),
  maxPrice = max(dataset$price), minPieces = min(dataset$pieces),
  maxPieces = max(dataset$pieces),
  themes = sort(unique(dataset$theme)),
  subthemes = sort(unique(dataset$subtheme)))
}
\arguments{
\item{dataset}{the dataset.}

\item{minYear}{the minimum year.}

\item{maxYear}{the maximum year.}

\item{minPrice}{the minimum price count.}

\item{maxPrice}{the maximum price count.}

\item{minPieces}{the minimum piece count.}

\item{maxPieces}{the maximum piece count.}

\item{themes}{the themes to include in the filtered dataset.}

\item{subthemes}{the subthemes to include in the filtered dataset.}
}
\value{
data.table 2 columns
}
\description{
Group by year to get total count of pieces.
}
