% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lego.R
\docType{package}
\name{lego}
\alias{lego}
\alias{lego-package}
\title{Lego set data 1970-2015}
\description{
Data for every lego set produced from 1970 to 2015 including number
of pieces per set, number of minifigures per set, and region specific MSRP.
}
\seealso{
\code{\link{legosets}}
}

