% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego-scrape_lego.R
\name{lego_exclusive_sets}
\alias{lego_exclusive_sets}
\title{Exclusive Lego Sets}
\usage{
lego_exclusive_sets(country_code = "en-ca")
}
\arguments{
\item{country_code}{A character string consisting of the country and language code. Defaults to \code{"en-ca"}.}
}
\value{
A dataframe of exclusive sets from \href{https://www.lego.com/en-ca/categories/exclusives}{Lego}.
}
\description{
Scrape currently available exclusive sets from \href{https://www.lego.com/en-ca/categories/exclusives}{Lego}.
}
\examples{
\dontrun{
lego_exclusive_sets(country_code = "en-ca")
}
}
