% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego-scrape_lego.R
\name{lego_theme_sets}
\alias{lego_theme_sets}
\title{Lego Sets for Specified Theme}
\usage{
lego_theme_sets(theme_url, country_code = "en-ca")
}
\arguments{
\item{theme_url}{A character string. URL path code format. See \code{theme_url} column returned by \code{current_lego_themes()} for valid options.}

\item{country_code}{A character string consisting of the country and language code. Defaults to \code{"en-ca"}.}
}
\value{
A dataframe of exclusive sets from \href{https://www.lego.com/en-ca/themes}{Lego}.
}
\description{
Scrape currently available sets for a specified theme from \href{https://www.lego.com/en-ca/themes}{Lego}.
}
\examples{
\dontrun{
lego_theme_sets(theme_url = "brickheadz")
}
}
