% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotLego}
\alias{plotLego}
\title{plot - wrapper for all of the interactive plot functionality. See the examples for usage & parameters.}
\usage{
plotLego(dataset, type)
}
\arguments{
\item{dataset}{the dataset.}

\item{type}{the type of ggplot to create - can be one of 'year', 'piece', 'price', 'theme', 'piece_ind', 'price_ind', piece_avg', piece_theme_avg'.}
}
\value{
the plot.
}
\description{
plot - wrapper for all of the interactive plot functionality. See the examples for usage & parameters.
}
\examples{
dataset <- read.lego(system.file('extdata', 'brickset-mysets-owned.csv', package ='lego'))
# Plot the number of sets by year.
plotLego(dataset, 'year')
Sys.sleep(1.25)
# Plot the number of pieces by year.
plotLego(dataset, 'piece')
Sys.sleep(1.25)
# Plot the price by year.
plotLego(dataset, 'price')
Sys.sleep(1.25)
# Plot the number of themes by year.
plotLego(dataset, 'theme')
Sys.sleep(1.25)
# Plot the number of pieces for each individual set by year.
plotLego(dataset, 'piece_ind')
Sys.sleep(1.25)
# Plot the price for each individual set by year.
plotLego(dataset, 'price_ind')
Sys.sleep(1.25)
# Plot the average number of pieces by year.
plotLego(dataset, 'piece_avg')
Sys.sleep(1.25)
# Plot the average number of pieces by theme.
plotLego(dataset, 'piece_theme_avg')
}
\keyword{lego}
