% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotPiecesByYear}
\alias{plotPiecesByYear}
\title{Plot number of pieces by year for each individual set.}
\usage{
plotPiecesByYear(dataset, dom = "piecesByYear", xAxisLabel = "Year",
  yAxisLabel = "Number of Pieces")
}
\arguments{
\item{dataset}{the dataset.}

\item{dom}{the dom.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - number of pieces.}
}
\value{
the piecesByYear plot.
}
\description{
Plot number of pieces by year for each individual set.
}
\examples{
dataset <- read.lego(system.file('extdata', 'brickset-mysets-owned.csv', package ='lego'))
plotPiecesByYear(filterByYearPricePieceTheme(dataset))
}
\keyword{lego}
