% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotPiecesByYearAvg}
\alias{plotPiecesByYearAvg}
\title{Plot the average number of pieces by year.}
\usage{
plotPiecesByYearAvg(dataset, dom = "piecesByYearAvg",
  xAxisLabel = "Year", yAxisLabel = "Number of Pieces")
}
\arguments{
\item{dataset}{the dataset.}

\item{dom}{the dom.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - number of pieces.}
}
\value{
the themesByYear plot.
}
\description{
Plot the average number of pieces by year.
}
