% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotPiecesCountByYear}
\alias{plotPiecesCountByYear}
\title{Plot number of pieces by year.}
\usage{
plotPiecesCountByYear(dataset, dom = "piecesCountByYear",
  xAxisLabel = "Year", yAxisLabel = "Number of Pieces")
}
\arguments{
\item{dataset}{the dataset.}

\item{dom}{the dom.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - the number of pieces.}
}
\value{
the piecesCountByYear plot.
}
\description{
Plot number of pieces by year.
}
