% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotPriceCountByYear}
\alias{plotPriceCountByYear}
\title{Plot price by year.}
\usage{
plotPriceCountByYear(dataset, dom = "priceCountByYear",
  xAxisLabel = "Year", yAxisLabel = "Price")
}
\arguments{
\item{dataset}{the dataset.}

\item{dom}{the dom.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - the price.}
}
\value{
the priceCountByYear plot.
}
\description{
Plot price by year.
}
