% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotSetsCountByYear}
\alias{plotSetsCountByYear}
\title{Plot number of sets by year}
\usage{
plotSetsCountByYear(dataset, dom = "setsByYear", xAxisLabel = "Year",
  yAxisLabel = "Number of Sets")
}
\arguments{
\item{dataset}{the dataset.}

\item{dom}{the dom.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - the number of sets.}
}
\value{
the setsByYear plot.
}
\description{
Plot number of sets by year
}
