% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotThemesCountByYear}
\alias{plotThemesCountByYear}
\title{Plot number of themes by year.}
\usage{
plotThemesCountByYear(dataset, dom = "themesByYear",
  xAxisLabel = "Year", yAxisLabel = "Number of Themes")
}
\arguments{
\item{dataset}{the dataset.}

\item{dom}{the dom.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - number of themes.}
}
\value{
the themesByYear plot.
}
\description{
Plot number of themes by year.
}
