% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read.lego}
\alias{read.lego}
\title{Read a brickset lego file.}
\usage{
read.lego(file)
}
\arguments{
\item{file}{the file.}
}
\value{
the lego brickset lego file.
}
\description{
Read a brickset lego file.
}
