% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{runShinyLego}
\alias{runShinyLego}
\title{Run the shiny application.}
\usage{
runShinyLego()
}
\description{
Run the shiny application.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  # run with Darren's default dataset
  runShinyLego()
  # supply your own dataset by saving it to the variable lego_data.
  lego_data <- read.lego(system.file('extdata', 'sets.csv', package ='lego'))
  runShinyLego()
}
}
\keyword{lego}
