% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego-scrape_lego.R
\name{scrape_lego_internal}
\alias{scrape_lego_internal}
\title{Internal Utility Function for Scraping Pages from \href{https://www.lego.com/en-ca/}{Lego}.}
\usage{
scrape_lego_internal(lego_url)
}
\arguments{
\item{lego_url}{A character string. URL to specific lego web page.}
}
\value{
A dataframe of data scraped from \href{https://www.lego.com/en-ca/}{Lego}.
}
\description{
Internal Utility Function for Scraping Pages from \href{https://www.lego.com/en-ca/}{Lego}.
}
\keyword{internal}
