% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{filterByYear}
\alias{filterByYear}
\title{Filter dataset only by year}
\usage{
filterByYear(dataset, minYear = min(dataset$year),
  maxYear = max(dataset$year))
}
\arguments{
\item{dataset}{the dataset.}

\item{minYear}{the minimum year. Defaults to be minimum year in the dataset.}

\item{maxYear}{the maximum year. Defaults to be maximum year in the dataset.}
}
\value{
the filtered dataset with the filters applied.
}
\description{
Filter dataset only by year
}
\examples{
dataset <- read.lego(system.file('extdata', 'brickset-mysets-owned.csv', package ='lego'))
filterByYear(dataset)
filterByYear(dataset, 2015)
filterByYear(dataset, maxYear=2016)

}
\keyword{lego}
