% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ggplotThemesCountByYear}
\alias{ggplotThemesCountByYear}
\title{Ggplot number of themes by year}
\usage{
ggplotThemesCountByYear(dataset, title = "Number of Themes per Year",
  xAxisLabel = "Year", yAxisLabel = "Number of Themes",
  caption = "Source: Darren Redmond's owned dataset")
}
\arguments{
\item{dataset}{the dataset.}

\item{title}{the title.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - the price.}

\item{caption}{the caption.}
}
\value{
the themesByYear plot.
}
\description{
Ggplot number of themes by year
}
\examples{
dataset <- read.lego(system.file('extdata', 'brickset-mysets-owned.csv', package ='lego'))
ggplotThemesCountByYear(groupByTheme(dataset))
ggplotThemesCountByYear(groupByTheme(dataset, themes=c('Star Wars')))
ggplotThemesCountByYear(groupByTheme(dataset, 2010, 2016, 5, 10, 100, 1000,
    c('Star Wars'), c('Episode I')))
ggplotThemesCountByYear(groupByTheme(dataset, 2014, 2015,
    themes=c('Star Wars'), subthemes=c('Episode I')))
}
\keyword{lego}
