% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{groupByPrice}
\alias{groupByPrice}
\title{Group by year to get total price.}
\usage{
groupByPrice(dataset, minYear = min(dataset$year),
  maxYear = max(dataset$year), minPrice = min(dataset$price),
  maxPrice = max(dataset$price), minPieces = min(dataset$pieces),
  maxPieces = max(dataset$pieces),
  themes = sort(unique(dataset$theme)),
  subthemes = sort(unique(dataset$subtheme)))
}
\arguments{
\item{dataset}{the dataset.}

\item{minYear}{the minimum year.}

\item{maxYear}{the maximum year.}

\item{minPrice}{the minimum price count.}

\item{maxPrice}{the maximum price count.}

\item{minPieces}{the minimum piece count.}

\item{maxPieces}{the maximum piece count.}

\item{themes}{the themes to include in the filtered dataset.}

\item{subthemes}{the subthemes to include in the filtered dataset.}
}
\value{
data.table 2 columns
}
\description{
Group by year to get total price.
}
\examples{
dataset <- read.lego(system.file('extdata', 'brickset-mysets-owned.csv', package ='lego'))
groupByPrice(dataset)
groupByPrice(dataset, themes=c('Star Wars'))
groupByPrice(dataset, 2010, 2016, 5, 10, 100, 1000, c('Star Wars'), c('Episode I'))
groupByPrice(dataset, 2014, 2015, themes=c('Star Wars'), subthemes=c('Episode I'))

}
\keyword{lego}
