% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{groupByYear}
\alias{groupByYear}
\title{Group by year and summarise by set.}
\usage{
groupByYear(dataset, minYear, maxYear, minPrice, maxPrice, minPieces,
  maxPieces, themes, subthemes)
}
\arguments{
\item{dataset}{the dataset.}

\item{minYear}{the minimum year.}

\item{maxYear}{the maximum year.}

\item{minPrice}{the minimum price count.}

\item{maxPrice}{the maximum price count.}

\item{minPieces}{the minimum piece count.}

\item{maxPieces}{the maximum piece count.}

\item{themes}{the themes to include in the filtered dataset.}

\item{subthemes}{the subthemes to include in the filtered dataset.}
}
\value{
the filtered dataset with the filters applied.
}
\description{
Group by year and summarise by set.
}
\keyword{lego}
