% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotPiecesByYear}
\alias{plotPiecesByYear}
\title{Plot number of pieces by year.}
\usage{
plotPiecesByYear(dataset, dom = "piecesByYear", xAxisLabel = "Year",
  yAxisLabel = "Number of Pieces")
}
\arguments{
\item{dataset}{the dataset.}

\item{dom}{the dom.}

\item{xAxisLabel}{the x axis label - the year.}

\item{yAxisLabel}{the y axis label - number of pieces.}
}
\value{
the piecesByYear plot.
}
\description{
Plot number of pieces by year.
}
