% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read.lego}
\alias{read.lego}
\title{Read a brickset lego file.}
\usage{
read.lego(file)
}
\arguments{
\item{file}{the file.}
}
\value{
the lego brickset lego file.
}
\description{
Read a brickset lego file.
}
\examples{
dataset <- read.lego(system.file('extdata', 'brickset-mysets-owned.csv', package ='lego'))
all_dataset <- read.lego(system.file('extdata', 'sets.csv', package ='lego'))
}
